Refactored distribution batch checking into distribution.py
